+++ src/application/context/goals/query/GoalBacklogPreviewItem.ts
@@ -1,0 +1,8 @@
+import { GoalStatusType } from "../../../../domain/goals/Constants.js";
+
+export interface GoalBacklogPreviewItem {
+  readonly goalId: string;
+  readonly title: string;
+  readonly status: GoalStatusType;
+  readonly createdAt: string;
+}